Domino implemented in C
In this project, I created a game called Domino. In this project, which I worked on with my friend, we made an interactive interface where you can choose what you want. You can save your game and play against the computer or with a friend. Inside the game, this interface shows your pieces for you to choose and play, and it tells you if you can’t play a piece and suggests that you buy one.